refactor(Neon): inline Operations.ts helpers into Project/Branch
Move waitForOperations + findProjectByName into Project.ts and
findBranchByName + the local createProjectBranch override into
Branch.ts. The Operations.ts module is gone.
